% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATScore_BriefIATDataset.R
\docType{data}
\name{BriefIAT}
\alias{BriefIAT}
\title{Sample Brief IAT Dataset (Abbreviated IAT)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140 rows and 6 columns.}
\usage{
data(BriefIAT)
}
\description{
This dataset contains sample data from a Brief IAT (a shorter, special form of the IAT).
This IAT has 140 trials, and this particular dataset would not meet criteria for exclusion.
}
\examples{
data(BriefIAT)
}
\keyword{datasets}
