% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATScore_SampleDataset.R
\docType{data}
\name{IAT}
\alias{IAT}
\title{Sample IAT Dataset (Typical)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 6 columns.}
\usage{
data(IAT)
}
\description{
This dataset contains sample data from a fairly typical IAT.
This IAT has 220 trials, and this particular dataset does not meet criteria for exclusion.
}
\examples{
data(IAT)
}
\keyword{datasets}
