% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IATScore_TooFastDataset.R
\docType{data}
\name{TooFastIAT}
\alias{TooFastIAT}
\title{Sample IAT Dataset (Participant went too fast)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 6 columns.
}
\usage{
data(TooFastIAT)
}
\description{
This dataset contains sample data from a participant who went too fast on the IAT.
This IAT has 220 trials, and this particular dataset would meet exclusion criteria.
}
\examples{
data(TooFastIAT)
}
\keyword{datasets}
